ml-engine/requirements.txt
flask==2.3.3
flask-cors==4.0.0
prophet==1.1.4
pandas==2.0.3
python-dotenv==1.0.0
uvicorn==0.23.2
fastapi==0.103.1